feat: CoreDataStack 타입 생성
- CoreData 구성요소들이 담긴 PersistentCloudeKitContainer를 CoreDataStack 클래스타입에 정의하였다.
