+++ CloudNotes/CloudNotes/Utility/CoreDataStack/CoreDataStack.swift
@@ -1,0 +1,52 @@
+//
+//  CoreDataStack.swift
+//  CloudNotes
+//
+//  Created by KimJaeYoun on 2021/09/09.
+//
+
+import CoreData
+
+final class CoreDataStack {
+    // MARK: Property
+    static let shared = CoreDataStack()
+    
+    lazy var persistentContainer: NSPersistentCloudKitContainer = {
+        let container = NSPersistentCloudKitContainer(name: "CloudNotes")
+        container.loadPersistentStores { (_, error) in
+            if let error = error as NSError? {
+                fatalError("Unresolved error \(error), \(error.userInfo)")
+            }
+        }
+        return container
+    }()
+    
+    lazy var managedContext: NSManagedObjectContext = {
+        return persistentContainer.viewContext
+    }()
+    
+    // MARK: initializer
+    private init() { }
+    
+    // MARK: Function
+    func saveContext () {
+        guard managedContext.hasChanges else {
+            return
+        }
+        
+        do {
+            try managedContext.save()
+        } catch {
+            let nserror = error as NSError
+            fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
+        }
+    }
+    
+    func makeFetchedResultsController(fetchRequest: NSFetchRequest<NSFetchRequestResult>, sectionNameKeyPath: String?, cacheName: String?) -> NSFetchedResultsController<NSFetchRequestResult> {
+        let controller = NSFetchedResultsController(fetchRequest: fetchRequest,
+                                                    managedObjectContext: managedContext,
+                                                    sectionNameKeyPath: sectionNameKeyPath,
+                                                    cacheName: cacheName)
+        return controller
+    }
+}
